requirements.txt
azure-identity
openai>=1.109.1
python-dotenv
pydantic
rich
dotenv-azd
aiohttp
faker
openai-agents
semantic-kernel
langgraph
langchain>=1.2.15
langchain_openai>=1.1.12
pydantic-ai==1.0.8
llama-index
llama-index-llms-openai-like
langchain-mcp-adapters
agent-framework-core==1.0.0
agent-framework-openai==1.0.0
agent-framework-devui>=1.0.0b260402
agent-framework-orchestrations>=1.0.0b260402
